Improve volume management

+++ .devcontainer/scripts/remove-volumes.sh
@@ -2,4 +2,10 @@
 set -e
 
 docker volume rm -f pgadmin pgdata vscode pgbootstrap
-docker volume rm -f $(docker volume ls -f name=fws --format json | jq -r .Name)
+
+FWS_VOLUME=$(docker volume ls -q -f "name=fws")
+
+if [ ! -z "$FWS_VOLUME" ]; then
+  docker volume rm -f $(docker volume ls -f name=fws --format json | jq -r .Name)
+fi
+
